Move RouteContexts to a contexts subdirectory
